requirements.txt
ipython
nslsii
sirepo-bluesky
